requirements.txt
Flask==2.2.2
Jinja2==3.0.2
mysql_connector_repackaged==0.3.1
numpy==1.21.2
python_bcrypt==0.3.2
spotipy==2.22.1
